I put plot skeletons and ideas for all of them, just need to put it into our code, and work on it.
For plot 4, the one with earthquakes, we need to first write it so that our dataset is used, and second we need to look at what the bar graph or maybe another plot should mean. In terms of interactive dashboard, perhaps we have this plot be the selector for all or maybe not?